+++ src/main/java/com/user/genericmod/init/ModItems.java
@@ -1,0 +1,14 @@
+package com.user.genericmod.init;
+
+import com.user.genericmod.item.*;
+
+/** 
+ *  All custom items are initialized and stored here. <br>
+ *  If you need an instance of an item, get it from here.
+ */
+public class ModItems 
+{
+	static
+	{
+	}
+}