+++ CHANGELOG.md
@@ -3,7 +3,12 @@ All notable changes to this mod project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and uses a custom versioning system.
 
-## Version 0.1.0
+## Version 0.2.0 - 2016-10-13
+### Added
+- Item and Block registration methods.
+- Container class for mod item instances.
+
+## Version 0.1.0 - 2016-10-13
 ### Added
 - A readme file.
 - A changelog file.